WIP hook useCheckout
